+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## XX.X.X [#1378](https://github.com/openfisca/openfisca-core/pull/1378)
+
+#### Technical changes
+
+- Allows calculation for big numbers
+    - Replace `int32 `with `int64 `and `float32 `with `float64`
+
 ## 44.7.0 [#1357](https://github.com/openfisca/openfisca-core/pull/1357)
 
 #### New features